requirements.txt
flask
pydub
Pillow
redis